refactor(jwt): categorize dtos, add def. role, fix some bugs
