+++ src/main/java/pl/studia/teletext/teletext_backend/api/admin/controllers/AuthController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import pl.studia.teletext.teletext_backend.api.admin.dtos.LoginRequest;
+import pl.studia.teletext.teletext_backend.api.admin.dtos.security.LoginRequest;
 import pl.studia.teletext.teletext_backend.domain.services.security.AuthService;
 
 @RestController
