--- src/main/java/pl/studia/teletext/teletext_backend/api/admin/dtos/LoginRequest.java
@@ -1,6 +1,0 @@
-package pl.studia.teletext.teletext_backend.api.admin.dtos;
-
-public record LoginRequest (
-  String username,
-  String password
-){ }
