+++ src/main/java/pl/studia/teletext/teletext_backend/api/admin/dtos/security/LoginResponse.java
@@ -1,4 +1,4 @@
-package pl.studia.teletext.teletext_backend.api.admin.dtos;
+package pl.studia.teletext.teletext_backend.api.admin.dtos.security;
 
 public record LoginResponse(
   String token,
