+++ src/main/java/pl/studia/teletext/teletext_backend/domain/services/TeletextPageStatsService.java
@@ -3,7 +3,7 @@ package pl.studia.teletext.teletext_backend.domain.services;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import pl.studia.teletext.teletext_backend.api.admin.dtos.TeletextPageStatsResponse;
+import pl.studia.teletext.teletext_backend.api.admin.dtos.stats.TeletextPageStatsResponse;
 import pl.studia.teletext.teletext_backend.api.admin.mappers.TeletextPageStatsMapper;
 import pl.studia.teletext.teletext_backend.exceptions.PageNotFoundException;
 import pl.studia.teletext.teletext_backend.domain.models.teletext.TeletextPageStats;
